adds global time model and updates components to use it
updates the time display, the current date display user log entries
counter, and user logs page to all use the new TimeModel. is a single
instance of the current time instead of having multiple throughout
app...fires a custom event when is a new day that other components can
listen for to know when to refresh themselves.